change pid paramteres from ints to floats

+++ Node2/src/pid.h
@@ -5,19 +5,19 @@
 
 
 typedef struct pid_param {
-    int16_t Kp;  /** Proportional gain */
-    int16_t Ki;  /** Intergral gain */
-    int16_t Kd;  /** Derivative gain */
-    int16_t T;   /** Time step of discretization in ms */
+    float Kp;  /** Proportional gain */
+    float Ki;  /** Intergral gain */
+    float Kd;  /** Derivative gain */
+    float T;   /** Time step of discretization in ms */
     
-    int16_t measurement;
-    int16_t setpoint;
+    float measurement;
+    float setpoint;
 
-    int16_t current_error;
-    int16_t previous_error;
-    int16_t cumulative_error;
+    float current_error;
+    float previous_error;
+    float cumulative_error;
 
-    int16_t output;
+    float output;
 } pid_t;
 
 
@@ -30,7 +30,7 @@ typedef struct pid_param {
  * @param kd Derivative gain
  * @param timestep Time step of discretization in ms in range [1, 250]
  */
-void pid_init(pid_t *pid, int16_t kp, int16_t ki, int16_t kd, int16_t timestep);
+void pid_init(pid_t *pid, float kp, float ki, float kd, float timestep);
 
 
 /**
